A bunch of changes

+++ db/migrate/20220601050029_create_line_items.rb
@@ -4,7 +4,7 @@ class CreateLineItems < ActiveRecord::Migration[7.0]
   def change
     create_table :line_items do |t|
       t.text :notes
-      t.decimal :sort_order
+      t.bigint :sort_order
       t.text :type_id
 
       t.timestamps
