+++ db/migrate/20220601050142_create_odometer_readings.rb
@@ -5,7 +5,7 @@ class CreateOdometerReadings < ActiveRecord::Migration[7.0]
     create_table :odometer_readings do |t|
       t.datetime :performed_at
       t.boolean :include_time
-      t.decimal :reading
+      t.bigint :reading
 
       t.timestamps
