+++ db/migrate/20220601050159_create_warranties.rb
@@ -3,8 +3,8 @@
 class CreateWarranties < ActiveRecord::Migration[7.0]
   def change
     create_table :warranties do |t|
-      t.decimal :miles
-      t.decimal :months
+      t.bigint :miles
+      t.bigint :months
       t.text :title
 
       t.timestamps
